Add kubelet build tag
