+++ pkg/collector/corechecks/embed/logs-agent.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2017 Datadog, Inc.
 
-// +build process
+// +build log
 
 package embed
