+++ pkg/tagger/collectors/kubelet_extract_test.go
@@ -3,43 +3,18 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2017 Datadog, Inc.
 
+// +build kubelet
+
 package collectors
 
 import (
 	"encoding/json"
 	"io/ioutil"
-	"sort"
 	"testing"
 
-	"github.com/stretchr/testify/require"
-
 	"github.com/DataDog/datadog-agent/pkg/util/kubernetes/kubelet"
+	"github.com/stretchr/testify/require"
 )
-
-func requireMatchInfo(t *testing.T, expected []*TagInfo, item *TagInfo) bool {
-	for _, template := range expected {
-		if template.Entity != item.Entity {
-			continue
-		}
-		if template.Source != item.Source {
-			continue
-		}
-		sort.Strings(template.LowCardTags)
-		sort.Strings(item.LowCardTags)
-		require.Equal(t, template.LowCardTags, item.LowCardTags)
-
-		sort.Strings(template.HighCardTags)
-		sort.Strings(item.HighCardTags)
-		require.Equal(t, template.HighCardTags, item.HighCardTags)
-
-		require.Equal(t, template.DeleteEntity, item.DeleteEntity)
-
-		return true
-	}
-
-	t.Logf("could not find expected result for entity %s with sourcce %s", item.Entity, item.Source)
-	return false
-}
 
 func TestKubeletPodTags(t *testing.T) {
 	raw, err := ioutil.ReadFile("./test/kubelet/podlist_1.6.json")
