+++ pkg/util/kubernetes/kubelet/types_kubelet.go
@@ -3,6 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2017 Datadog, Inc.
 
+// +build kubelet
+
 package kubelet
 
 // Pod contains fields for unmarshalling a Pod
